// types/models/box.ts
export type Box = {
  id: number;
  cust_code: number | string | null;
  cust_name: string;
  cust_name_e: string;
  address: string | null;
  address_e: string | null;
  city_id: number | string | null;
  phone: string | null;
  phone2: string | null;
  fax: string | null;
  mobile: string | null;
  email: string | null;
  post: string | null;
  vat_no: string | number | null;
  cr_no: string | number | null;
  gov: string | null;
  city: string | null;
  area: string | null;
  street: string | null;
  build_no: string | null;
  post_no: string | null;
  post_code: string | null;
  col_id: number | string | null;
  debt_ceiling: string | number | null;
  max_vouch: string | number | null;
  max_inv: string | number | null;
  max_vouch_date: string | null;
  max_inv_date: string | null;
  hide: boolean;
  handling: string | null;
  handling_e: string | null;
  action: string | null;
  acc2: number | string | null;
  perc: string | number | null;
  box_type: number | string | null;
  expt: boolean;
  ordr: number | null;
  cat_id: number | string | null;
  cust_status: number | null;
  cr_date: string | null;
  cr_user: string | null;
  upd_date: string | null;
  upd_user: string | null;
  com: number | string | null;
  cust_type: number | null;
  cur: number | string | null;
  acc: number | string | null;
};

export type GetBoxesParams = {
  xcom_id?: number | string;
};